requirements.txt
numpy==1.19.2
matplotlib==3.4.0
seaborn==0.11.1
pandas==1.2.3
Django==3.1.7
kmodes==0.11.0
djangorestframework_simplejwt==4.7.1
djangorestframework==3.12.4
